[Refactor] Unexported logic utils function for working code
